% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rheap.R
\name{dnheap}
\alias{dnheap}
\alias{dnheap.minheap}
\alias{dnheap.maxheap}
\title{Heapify down from position}
\usage{
dnheap(heap, pos, size)

\method{dnheap}{minheap}(heap, pos, size)

\method{dnheap}{maxheap}(heap, pos, size)
}
\arguments{
\item{heap}{Heap object}

\item{pos}{Position to heapify down from}

\item{size}{Size of heap to consider}
}
\value{
Updated heap object
}
\description{
Maintain heap property downwards from position
}
