% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rheap.R
\name{head}
\alias{head}
\title{Extract top element from heap}
\usage{
head(heap)
}
\arguments{
\item{heap}{Heap object}
}
\value{
Head element of heap
}
\description{
Extract the top (min or max) element from a heap
}
