% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rheap.R
\name{insert}
\alias{insert}
\alias{insert.minheap}
\alias{insert.maxheap}
\title{Insert into fixed-size heap}
\usage{
insert(heap, value)

\method{insert}{minheap}(heap, value)

\method{insert}{maxheap}(heap, value)
}
\arguments{
\item{heap}{Heap object}

\item{value}{Numeric value to insert}
}
\value{
Updated heap object
}
\description{
Insert element into fixed-size heap
}
\details{
For a min-heap, finds the maximum leaf element and replaces it with the new value
if the new value is smaller. For a max-heap, finds the minimum leaf element and
replaces it with the new value if the new value is larger.
}
