% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rheap.R
\name{maxheap}
\alias{maxheap}
\title{Create a max-heap}
\usage{
maxheap(x = numeric(0))
}
\arguments{
\item{x}{Numeric vector to convert to heap}
}
\value{
A maxheap object
}
\description{
Create a max-heap from a vector using Floyd's algorithm
}
