% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rheap.R
\name{minheap}
\alias{minheap}
\title{Create a min-heap}
\usage{
minheap(x = numeric(0))
}
\arguments{
\item{x}{Numeric vector to convert to heap}
}
\value{
A minheap object
}
\description{
Create a min-heap from a vector using Floyd's algorithm
}
