% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rheap.R
\name{npoppush}
\alias{npoppush}
\alias{npoppush.minheap}
\alias{npoppush.maxheap}
\title{Repeatedly poppush elements into a heap}
\usage{
npoppush(heap, values)

\method{npoppush}{minheap}(heap, values)

\method{npoppush}{maxheap}(heap, values)
}
\arguments{
\item{heap}{Heap object}

\item{values}{Numeric vector of values to push}
}
\value{
Updated heap object
}
\description{
Repeatedly pop then push into a heap.
}
\details{
For each value in values, the top element (minimum or maximum) of the heap
is popped and the new value is pushed to the heap. The popped values are
discarded and the resulting heap is returned.
}
