% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rheap.R
\name{npushpop}
\alias{npushpop}
\alias{npushpop.minheap}
\alias{npushpop.maxheap}
\title{Repeatedly Push and Pop a Heap}
\usage{
npushpop(heap, values)

\method{npushpop}{minheap}(heap, values)

\method{npushpop}{maxheap}(heap, values)
}
\arguments{
\item{heap}{Heap object}

\item{values}{Numeric vector of values to push}
}
\value{
Updated heap object
}
\description{
Repeatedly push then pop a heap
}
\details{
For each value in values, the value is pushed to the heap and then the top
(minimum or maximum) of the resulting heap is popped.  The popped values are
discarded and the resulting heap is returned.
}
