% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rheap.R
\name{pop}
\alias{pop}
\alias{pop.minheap}
\alias{pop.maxheap}
\title{Pop elements from a heap}
\usage{
pop(heap, n = 1L)

\method{pop}{minheap}(heap, n = 1L)

\method{pop}{maxheap}(heap, n = 1L)
}
\arguments{
\item{heap}{Heap object}

\item{n}{Number of elements to pop (default: 1)}
}
\value{
List containing vector of popped values and updated heap
}
\description{
Pop multiple elements from a heap
}
