% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rheap.R
\name{poppush}
\alias{poppush}
\alias{poppush.minheap}
\alias{poppush.maxheap}
\title{Pop top element then push new element}
\usage{
poppush(heap, value)

\method{poppush}{minheap}(heap, value)

\method{poppush}{maxheap}(heap, value)
}
\arguments{
\item{heap}{Heap object}

\item{value}{Numeric value to push}
}
\value{
List containing the top element and updated heap
}
\description{
Pop the top element then push a new element into a heap
}
\details{
This function is more efficient than separate calls to pop and push.
}
