% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rheap.R
\name{print.maxheap}
\alias{print.maxheap}
\title{Print a max-heap}
\usage{
\method{print}{maxheap}(x, ...)
}
\arguments{
\item{x}{Max-heap object}

\item{...}{currently ignored}
}
\description{
Print a max-heap
}
