% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rheap.R
\name{print.minheap}
\alias{print.minheap}
\title{Print a min-heap}
\usage{
\method{print}{minheap}(x, ...)
}
\arguments{
\item{x}{Min-heap object}

\item{...}{currently ignored}
}
\description{
Print a min-heap
}
