% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rheap.R
\name{push}
\alias{push}
\alias{push.minheap}
\alias{push.maxheap}
\title{Push element onto heap}
\usage{
push(heap, value)

\method{push}{minheap}(heap, value)

\method{push}{maxheap}(heap, value)
}
\arguments{
\item{heap}{Heap object}

\item{value}{Numeric value to push}
}
\value{
Updated heap object
}
\description{
Push an element onto a heap
}
