% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rheap.R
\name{pushpop}
\alias{pushpop}
\alias{pushpop.minheap}
\alias{pushpop.maxheap}
\title{Push and pop top element}
\usage{
pushpop(heap, value)

\method{pushpop}{minheap}(heap, value)

\method{pushpop}{maxheap}(heap, value)
}
\arguments{
\item{heap}{Heap object}

\item{value}{Numeric value to push}
}
\value{
List containing the top element and updated heap
}
\description{
Push a new element into a heap then pop the top element
}
\details{
This function is more efficient than separate calls to push and pop.
}
