% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rheap-package.R
\docType{package}
\name{Rheap-package}
\alias{Rheap}
\alias{Rheap-package}
\title{Rheap: Min and Max Heaps}
\description{
Provides min and max heaps for R.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/SWotherspoon/Rheap}
  \item Report bugs at \url{https://github.com/SWotherspoon/Rheap/issues}
}

}
\author{
\strong{Maintainer}: Simon Wotherspoon \email{Simon.Wotherspoon@aad.gov.au} (\href{https://orcid.org/0000-0002-6947-4445}{ORCID})

}
\keyword{internal}
