% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rheap-package.R
\docType{package}
\name{Rheap}
\alias{Rheap-package}
\alias{Rheap}
\title{R Implementation of Min- and Max-Heaps}
\description{
Provides implementations of min-heap and max-heaps.
}
\author{
\strong{Maintainer}: Simon Wotherspoon \email{Simon.Wotherspoon@aad.gov.au} (\href{https://orcid.org/0000-0002-6947-4445}{ORCID})

}
