% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rheap.R
\name{upheap}
\alias{upheap}
\alias{upheap.minheap}
\alias{upheap.maxheap}
\title{Heapify up from position}
\usage{
upheap(heap, pos)

\method{upheap}{minheap}(heap, pos)

\method{upheap}{maxheap}(heap, pos)
}
\arguments{
\item{heap}{Heap object}

\item{pos}{Position to heapify up from}
}
\value{
Updated heap object
}
\description{
Maintain heap property upwards from position
}
