% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rheap.R
\name{pop}
\alias{pop}
\alias{pop.minheap}
\alias{pop.maxheap}
\title{Pop top element from a heap}
\usage{
pop(heap)

\method{pop}{minheap}(heap)

\method{pop}{maxheap}(heap)
}
\arguments{
\item{heap}{Heap object}
}
\value{
List containing value and updated heap
}
\description{
Pop top element from a heap
}
