% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rheap.R
\name{push}
\alias{push}
\alias{push.minheap}
\alias{push.maxheap}
\title{Push elements onto heap}
\usage{
push(heap, values)

\method{push}{minheap}(heap, values)

\method{push}{maxheap}(heap, values)
}
\arguments{
\item{heap}{Heap object}

\item{values}{Numeric vector of values to push}
}
\value{
Updated heap object
}
\description{
Push one or more elements onto a heap
}
